pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "math-prm-verifier"
version = "0.1.0"
description = "A Process Reward Model (PRM) for Math Reasoning using Qwen-2.5 and System 2 Search."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy<2.0.0",
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "datasets>=2.14.0",
    "accelerate>=0.25.0",
    "trl>=0.7.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "vllm>=0.2.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
